Struct cntk::DataMap

pub struct DataMap { /* fields omitted */ }

Wrapper around unordered_map to pass bindings to Function evaluation

Methods

impl DataMap

fn new() -> DataMap

Creates an empty DataMap

fn add(&mut self, variable: &Variable, value: &Value)

Adds binding to DataMap. If mapping for given Variable exists, it will be overwritten.

fn add_null(&mut self, variable: &Variable)

Adds binding to null to DataMap. Useful, when we want function evaluation to create the Value.

fn get(&self, variable: &Variable) -> Option<Value>